bksys_gateway_client/pyproject.toml
[tool.poetry]
name = "bksys_gateway_client"
version = "0.1.0"
description = ""
authors = ["malo.girard <malo.girard@jellysmack.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
astroid = "^3.0.1"
aiohttp = "^3.8.6"
circuitbreaker = "^2.0.0"
prometheus-fastapi-instrumentator = "^6.1.0"
python-logging-loki = "^0.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
